refactored solvers out of their experimental folders and into mainline, added Direction Indicator documentation
