--- Assets/MRTK/Examples/Experimental/Solvers.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 9607ce05e5b18704c868c6e9e17d8864
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
