+++ Assets/MRTK/SDK/Editor/Inspectors/Utilities/Solvers/FollowInspector.cs
@@ -1,12 +1,11 @@
 ﻿// Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-using Microsoft.MixedReality.Toolkit.Experimental.Utilities;
-using Microsoft.MixedReality.Toolkit.Utilities.Editor.Solvers;
+using Microsoft.MixedReality.Toolkit.Utilities.Solvers;
 using UnityEditor;
 using UnityEngine;
 
-namespace Microsoft.MixedReality.Toolkit.Experimental.Editor
+namespace Microsoft.MixedReality.Toolkit.Utilities.Editor.Solvers
 {
     [CustomEditor(typeof(Follow))]
     [CanEditMultipleObjects]
