+++ Assets/MRTK/SDK/Features/Utilities/Solvers/DirectionalIndicator.cs
@@ -1,11 +1,9 @@
 ﻿// Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-using Microsoft.MixedReality.Toolkit.Utilities;
-using Microsoft.MixedReality.Toolkit.Utilities.Solvers;
 using UnityEngine;
 
-namespace Microsoft.MixedReality.Toolkit.Experimental.Utilities
+namespace Microsoft.MixedReality.Toolkit.Utilities.Solvers
 {
     /// <summary>
     /// This solver determines the position and orientation of an object as a directional indicator. 
@@ -23,14 +21,14 @@ namespace Microsoft.MixedReality.Toolkit.Experimental.Utilities
         public Transform DirectionalTarget;
 
         /// <summary>
-        /// The minimum scale for this the indicator object
+        /// The minimum scale for the indicator object
         /// </summary>
         [Tooltip("The offset from center to place the indicator. If frame of reference is the camera, then the object will be this distance from center of screen")]
         [Min(0.0f)]
         public float MinIndicatorScale = 0.05f;
 
         /// <summary>
-        /// The maximum scale for this the indicator object
+        /// The maximum scale for the indicator object
         /// </summary>
         [Tooltip("The offset from center to place the indicator. If frame of reference is the camera, then the object will be this distance from center of screen")]
         [Min(0.0f)]
